+++ packages/expo-updates/android/src/androidTest/java/expo/modules/updates/loader/FileDownloaderManifestParsingTest.kt
@@ -44,15 +44,15 @@ class FileDownloaderManifestParsingTest {
     var errorOccurred = false
     var resultUpdateManifest: UpdateManifest? = null
 
-    FileDownloader(context).parseManifestResponse(
+    FileDownloader(context).parseRemoteUpdateResponse(
       response, configuration,
-      object : FileDownloader.ManifestDownloadCallback {
+      object : FileDownloader.RemoteUpdateDownloadCallback {
         override fun onFailure(message: String, e: Exception) {
           errorOccurred = true
         }
 
-        override fun onSuccess(updateManifest: UpdateManifest) {
-          resultUpdateManifest = updateManifest
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateManifest = updateResponse.manifestUpdateResponsePart?.updateManifest
         }
       }
     )
@@ -69,6 +69,7 @@ class FileDownloaderManifestParsingTest {
     val contentType = "multipart/mixed; boundary=$boundary"
 
     val extensions = "{}"
+    val directive = CertificateFixtures.testDirectiveNoUpdateAvailable
 
     val multipartBody = MultipartBody.Builder(boundary)
       .setType(MultipartBody.MIXED)
@@ -83,6 +84,10 @@ class FileDownloaderManifestParsingTest {
       .addPart(
         mapOf("Content-Disposition" to "form-data; name=\"extensions\"; filename=\"hello3\"").toHeaders(),
         RequestBody.create("application/json; charset=utf-8".toMediaTypeOrNull(), extensions)
+      )
+      .addPart(
+        mapOf("Content-Disposition" to "form-data; name=\"directive\"; filename=\"hello3\"").toHeaders(),
+        RequestBody.create("application/json; charset=utf-8".toMediaTypeOrNull(), directive)
       )
       .build()
 
@@ -102,24 +107,189 @@ class FileDownloaderManifestParsingTest {
     )
 
     var errorOccurred = false
-    var resultUpdateManifest: UpdateManifest? = null
+    var resultUpdateResponse: UpdateResponse? = null
 
-    FileDownloader(context).parseManifestResponse(
+    FileDownloader(context).parseRemoteUpdateResponse(
       response, configuration,
-      object : FileDownloader.ManifestDownloadCallback {
+      object : FileDownloader.RemoteUpdateDownloadCallback {
         override fun onFailure(message: String, e: Exception) {
           errorOccurred = true
         }
 
-        override fun onSuccess(updateManifest: UpdateManifest) {
-          resultUpdateManifest = updateManifest
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateResponse = updateResponse
         }
       }
     )
 
     Assert.assertFalse(errorOccurred)
-    Assert.assertNotNull(resultUpdateManifest)
-    Assert.assertFalse(resultUpdateManifest!!.manifest.isVerified())
+
+    Assert.assertNotNull(resultUpdateResponse)
+    Assert.assertNotNull(resultUpdateResponse!!.manifestUpdateResponsePart)
+    Assert.assertFalse(resultUpdateResponse!!.manifestUpdateResponsePart!!.updateManifest.manifest.isVerified())
+
+    Assert.assertNotNull(resultUpdateResponse!!.directiveUpdateResponsePart)
+    Assert.assertTrue(resultUpdateResponse!!.directiveUpdateResponsePart!!.updateDirective is UpdateDirective.NoUpdateAvailableUpdateDirective)
+  }
+
+  @Test
+  fun testManifestParsing_MultipartBodyOnlyDirective() {
+    val context = InstrumentationRegistry.getInstrumentation().targetContext
+    val boundary = "blah"
+    val contentType = "multipart/mixed; boundary=$boundary"
+
+    val directive = CertificateFixtures.testDirectiveNoUpdateAvailable
+
+    val multipartBody = MultipartBody.Builder(boundary)
+      .setType(MultipartBody.MIXED)
+      .addPart(
+        mapOf("Content-Disposition" to "form-data; name=\"directive\"; filename=\"hello3\"").toHeaders(),
+        RequestBody.create("application/json; charset=utf-8".toMediaTypeOrNull(), directive)
+      )
+      .build()
+
+    val contentBuffer = Buffer().also { multipartBody.writeTo(it) }
+
+    val response = mockk<Response>().apply {
+      every { header("content-type") } returns contentType
+      every { headers } returns mapOf("content-type" to contentType).toHeaders()
+      every { body } returns ResponseBody.create(MultipartBody.MIXED, contentBuffer.readByteArray())
+    }
+
+    val configuration = UpdatesConfiguration(
+      null,
+      mapOf(
+        UpdatesConfiguration.UPDATES_CONFIGURATION_UPDATE_URL_KEY to Uri.parse("https://exp.host/@test/test"),
+      )
+    )
+
+    var errorOccurred = false
+    var resultUpdateResponse: UpdateResponse? = null
+
+    FileDownloader(context).parseRemoteUpdateResponse(
+      response, configuration,
+      object : FileDownloader.RemoteUpdateDownloadCallback {
+        override fun onFailure(message: String, e: Exception) {
+          errorOccurred = true
+        }
+
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateResponse = updateResponse
+        }
+      }
+    )
+
+    Assert.assertFalse(errorOccurred)
+
+    Assert.assertNotNull(resultUpdateResponse)
+    Assert.assertNull(resultUpdateResponse!!.manifestUpdateResponsePart)
+
+    Assert.assertNotNull(resultUpdateResponse!!.directiveUpdateResponsePart)
+    Assert.assertTrue(resultUpdateResponse!!.directiveUpdateResponsePart!!.updateDirective is UpdateDirective.NoUpdateAvailableUpdateDirective)
+  }
+
+  @Test
+  fun testManifestParsing_MultipartBodyOnlyDirective_v0CompatibilityMode() {
+    val context = InstrumentationRegistry.getInstrumentation().targetContext
+    val boundary = "blah"
+    val contentType = "multipart/mixed; boundary=$boundary"
+
+    val directive = CertificateFixtures.testDirectiveNoUpdateAvailable
+
+    val multipartBody = MultipartBody.Builder(boundary)
+      .setType(MultipartBody.MIXED)
+      .addPart(
+        mapOf("Content-Disposition" to "form-data; name=\"directive\"; filename=\"hello3\"").toHeaders(),
+        RequestBody.create("application/json; charset=utf-8".toMediaTypeOrNull(), directive)
+      )
+      .build()
+
+    val contentBuffer = Buffer().also { multipartBody.writeTo(it) }
+
+    val response = mockk<Response>().apply {
+      every { header("content-type") } returns contentType
+      every { headers } returns mapOf("content-type" to contentType).toHeaders()
+      every { body } returns ResponseBody.create(MultipartBody.MIXED, contentBuffer.readByteArray())
+    }
+
+    val configuration = UpdatesConfiguration(
+      null,
+      mapOf(
+        UpdatesConfiguration.UPDATES_CONFIGURATION_UPDATE_URL_KEY to Uri.parse("https://exp.host/@test/test"),
+        UpdatesConfiguration.UPDATES_CONFIGURATION_ENABLE_EXPO_UPDATES_PROTOCOL_V0_COMPATIBILITY_MODE to true,
+      )
+    )
+
+    var errorOccurred: Exception? = null
+    var resultUpdateManifest: UpdateManifest? = null
+
+    FileDownloader(context).parseRemoteUpdateResponse(
+      response, configuration,
+      object : FileDownloader.RemoteUpdateDownloadCallback {
+        override fun onFailure(message: String, e: Exception) {
+          errorOccurred = e
+        }
+
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateManifest = updateResponse.manifestUpdateResponsePart?.updateManifest
+        }
+      }
+    )
+
+    Assert.assertEquals("Multipart response missing manifest part. Manifest is required in version 0 of the expo-updates protocol. This may be due to the update being a rollback or other directive.", errorOccurred!!.message)
+    Assert.assertNull(resultUpdateManifest)
+  }
+
+  @Test
+  fun testManifestParsing_MultipartBodyNoRelevantParts() {
+    val context = InstrumentationRegistry.getInstrumentation().targetContext
+    val boundary = "blah"
+    val contentType = "multipart/mixed; boundary=$boundary"
+
+    val multipartBody = MultipartBody.Builder(boundary)
+      .setType(MultipartBody.MIXED)
+      .addPart(
+        mapOf("Content-Disposition" to "form-data; name=\"fake\"; filename=\"hello3\"").toHeaders(),
+        RequestBody.create("application/json; charset=utf-8".toMediaTypeOrNull(), "")
+      )
+      .build()
+
+    val contentBuffer = Buffer().also { multipartBody.writeTo(it) }
+
+    val response = mockk<Response>().apply {
+      every { header("content-type") } returns contentType
+      every { headers } returns mapOf("content-type" to contentType).toHeaders()
+      every { body } returns ResponseBody.create(MultipartBody.MIXED, contentBuffer.readByteArray())
+    }
+
+    val configuration = UpdatesConfiguration(
+      null,
+      mapOf(
+        UpdatesConfiguration.UPDATES_CONFIGURATION_UPDATE_URL_KEY to Uri.parse("https://exp.host/@test/test"),
+      )
+    )
+
+    var errorOccurred = false
+    var resultUpdateResponse: UpdateResponse? = null
+
+    FileDownloader(context).parseRemoteUpdateResponse(
+      response, configuration,
+      object : FileDownloader.RemoteUpdateDownloadCallback {
+        override fun onFailure(message: String, e: Exception) {
+          errorOccurred = true
+        }
+
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateResponse = updateResponse
+        }
+      }
+    )
+
+    Assert.assertFalse(errorOccurred)
+
+    Assert.assertNotNull(resultUpdateResponse)
+    Assert.assertNull(resultUpdateResponse!!.manifestUpdateResponsePart)
+    Assert.assertNull(resultUpdateResponse!!.directiveUpdateResponsePart)
   }
 
   @Test
@@ -155,15 +325,15 @@ class FileDownloaderManifestParsingTest {
     var errorOccurred = false
     var resultUpdateManifest: UpdateManifest? = null
 
-    FileDownloader(context).parseManifestResponse(
+    FileDownloader(context).parseRemoteUpdateResponse(
       response, configuration,
-      object : FileDownloader.ManifestDownloadCallback {
+      object : FileDownloader.RemoteUpdateDownloadCallback {
         override fun onFailure(message: String, e: Exception) {
           errorOccurred = true
         }
 
-        override fun onSuccess(updateManifest: UpdateManifest) {
-          resultUpdateManifest = updateManifest
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateManifest = updateResponse.manifestUpdateResponsePart?.updateManifest
         }
       }
     )
@@ -186,6 +356,7 @@ class FileDownloaderManifestParsingTest {
     )
 
     val extensions = "{}"
+    val directive = CertificateFixtures.testDirectiveNoUpdateAvailable
 
     val multipartBody = MultipartBody.Builder(boundary)
       .setType(MultipartBody.MIXED)
@@ -204,6 +375,14 @@ class FileDownloaderManifestParsingTest {
       .addPart(
         mapOf("Content-Disposition" to "form-data; name=\"extensions\"; filename=\"hello3\"").toHeaders(),
         RequestBody.create("application/json; charset=utf-8".toMediaTypeOrNull(), extensions)
+      )
+      .addPart(
+        mapOf(
+          "Content-Disposition" to "form-data; name=\"directive\"; filename=\"hello3\"",
+          "expo-signature" to CertificateFixtures.testDirectiveNoUpdateAvailableSignature
+        )
+          .toHeaders(),
+        RequestBody.create("application/json; charset=utf-8".toMediaTypeOrNull(), directive)
       )
       .build()
 
@@ -228,24 +407,28 @@ class FileDownloaderManifestParsingTest {
     )
 
     var errorOccurred = false
-    var resultUpdateManifest: UpdateManifest? = null
+    var resultUpdateResponse: UpdateResponse? = null
 
-    FileDownloader(context).parseManifestResponse(
+    FileDownloader(context).parseRemoteUpdateResponse(
       response, configuration,
-      object : FileDownloader.ManifestDownloadCallback {
+      object : FileDownloader.RemoteUpdateDownloadCallback {
         override fun onFailure(message: String, e: Exception) {
           errorOccurred = true
         }
 
-        override fun onSuccess(updateManifest: UpdateManifest) {
-          resultUpdateManifest = updateManifest
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateResponse = updateResponse
         }
       }
     )
 
     Assert.assertFalse(errorOccurred)
-    Assert.assertNotNull(resultUpdateManifest)
-    Assert.assertTrue(resultUpdateManifest!!.manifest.isVerified())
+
+    Assert.assertNotNull(resultUpdateResponse)
+    Assert.assertTrue(resultUpdateResponse!!.manifestUpdateResponsePart!!.updateManifest.manifest.isVerified())
+
+    Assert.assertNotNull(resultUpdateResponse!!.directiveUpdateResponsePart)
+    Assert.assertTrue(resultUpdateResponse!!.directiveUpdateResponsePart!!.updateDirective is UpdateDirective.NoUpdateAvailableUpdateDirective)
   }
 
   @Test
@@ -280,15 +463,15 @@ class FileDownloaderManifestParsingTest {
     var errorOccurred: Exception? = null
     var resultUpdateManifest: UpdateManifest? = null
 
-    FileDownloader(context).parseManifestResponse(
+    FileDownloader(context).parseRemoteUpdateResponse(
       response, configuration,
-      object : FileDownloader.ManifestDownloadCallback {
+      object : FileDownloader.RemoteUpdateDownloadCallback {
         override fun onFailure(message: String, e: Exception) {
           errorOccurred = e
         }
 
-        override fun onSuccess(updateManifest: UpdateManifest) {
-          resultUpdateManifest = updateManifest
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateManifest = updateResponse.manifestUpdateResponsePart?.updateManifest
         }
       }
     )
@@ -310,6 +493,7 @@ class FileDownloaderManifestParsingTest {
     )
 
     val extensions = "{}"
+    val directive = CertificateFixtures.testDirectiveNoUpdateAvailable
 
     val leafCert = getTestCertificate(TestCertificateType.CHAIN_LEAF)
     val intermediateCert = getTestCertificate(TestCertificateType.CHAIN_INTERMEDIATE)
@@ -337,6 +521,14 @@ class FileDownloaderManifestParsingTest {
         mapOf("Content-Disposition" to "form-data; name=\"certificate_chain\"; filename=\"hello4\"").toHeaders(),
         RequestBody.create("application/x-pem-file; charset=utf-8".toMediaTypeOrNull(), leafCert + intermediateCert)
       )
+      .addPart(
+        mapOf(
+          "Content-Disposition" to "form-data; name=\"directive\"; filename=\"hello3\"",
+          "expo-signature" to CertificateFixtures.testDirectiveNoUpdateAvailableValidChainLeafSignature
+        )
+          .toHeaders(),
+        RequestBody.create("application/json; charset=utf-8".toMediaTypeOrNull(), directive)
+      )
       .build()
 
     val contentBuffer = Buffer().also { multipartBody.writeTo(it) }
@@ -362,24 +554,27 @@ class FileDownloaderManifestParsingTest {
     )
 
     var errorOccurred = false
-    var resultUpdateManifest: UpdateManifest? = null
+    var resultUpdateResponse: UpdateResponse? = null
 
-    FileDownloader(context).parseManifestResponse(
+    FileDownloader(context).parseRemoteUpdateResponse(
       response, configuration,
-      object : FileDownloader.ManifestDownloadCallback {
+      object : FileDownloader.RemoteUpdateDownloadCallback {
         override fun onFailure(message: String, e: Exception) {
           errorOccurred = true
         }
 
-        override fun onSuccess(updateManifest: UpdateManifest) {
-          resultUpdateManifest = updateManifest
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateResponse = updateResponse
         }
       }
     )
 
     Assert.assertFalse(errorOccurred)
-    Assert.assertNotNull(resultUpdateManifest)
-    Assert.assertTrue(resultUpdateManifest!!.manifest.isVerified())
+    Assert.assertNotNull(resultUpdateResponse!!.manifestUpdateResponsePart?.updateManifest)
+    Assert.assertTrue(resultUpdateResponse!!.manifestUpdateResponsePart?.updateManifest!!.manifest.isVerified())
+
+    Assert.assertNotNull(resultUpdateResponse!!.directiveUpdateResponsePart)
+    Assert.assertTrue(resultUpdateResponse!!.directiveUpdateResponsePart!!.updateDirective is UpdateDirective.NoUpdateAvailableUpdateDirective)
   }
 
   @Test
@@ -449,21 +644,101 @@ class FileDownloaderManifestParsingTest {
     var errorOccurred: Exception? = null
     var resultUpdateManifest: UpdateManifest? = null
 
-    FileDownloader(context).parseManifestResponse(
+    FileDownloader(context).parseRemoteUpdateResponse(
       response, configuration,
-      object : FileDownloader.ManifestDownloadCallback {
+      object : FileDownloader.RemoteUpdateDownloadCallback {
         override fun onFailure(message: String, e: Exception) {
           errorOccurred = e
         }
 
-        override fun onSuccess(updateManifest: UpdateManifest) {
-          resultUpdateManifest = updateManifest
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateManifest = updateResponse.manifestUpdateResponsePart?.updateManifest
         }
       }
     )
 
     Assert.assertEquals("Invalid certificate for manifest project ID or scope key", errorOccurred!!.message)
     Assert.assertNull(resultUpdateManifest)
+  }
+
+  @Test
+  fun testManifestParsing_MultipartBodySignedCertificateParticularExperience_IncorrectExperienceInDirective() {
+    val context = InstrumentationRegistry.getInstrumentation().targetContext
+    val boundary = "blah"
+    val contentType = "multipart/mixed; boundary=$boundary"
+
+    val headersMap = mapOf(
+      "expo-protocol-version" to "0",
+      "expo-sfv-version" to "0",
+      "content-type" to contentType,
+    )
+
+    val directive = CertificateFixtures.testDirectiveNoUpdateAvailableIncorrectProjectId
+
+    val leafCert = getTestCertificate(TestCertificateType.CHAIN_LEAF)
+    val intermediateCert = getTestCertificate(TestCertificateType.CHAIN_INTERMEDIATE)
+    val rootCert = getTestCertificate(TestCertificateType.CHAIN_ROOT)
+
+    val multipartBody = MultipartBody.Builder(boundary)
+      .setType(MultipartBody.MIXED)
+      .addPart(
+        mapOf("Content-Disposition" to "form-data; name=\"extraneous\"; filename=\"hello1\"").toHeaders(),
+        RequestBody.create("text/plain; charset=utf-8".toMediaTypeOrNull(), "hello")
+      )
+      .addPart(
+        mapOf("Content-Disposition" to "form-data; name=\"certificate_chain\"; filename=\"hello4\"").toHeaders(),
+        RequestBody.create("application/x-pem-file; charset=utf-8".toMediaTypeOrNull(), leafCert + intermediateCert)
+      )
+      .addPart(
+        mapOf(
+          "Content-Disposition" to "form-data; name=\"directive\"; filename=\"hello3\"",
+          "expo-signature" to CertificateFixtures.testDirectiveNoUpdateAvailableValidChainLeafSignatureIncorrectProjectId
+        )
+          .toHeaders(),
+        RequestBody.create("application/json; charset=utf-8".toMediaTypeOrNull(), directive)
+      )
+      .build()
+
+    val contentBuffer = Buffer().also { multipartBody.writeTo(it) }
+
+    val response = mockk<Response>().apply {
+      headersMap.forEach {
+        every { header(it.key) } returns it.value
+      }
+      every { headers } returns headersMap.toHeaders()
+      every { body } returns ResponseBody.create(MultipartBody.MIXED, contentBuffer.readByteArray())
+    }
+
+    val configuration = UpdatesConfiguration(
+      null,
+      mapOf(
+        UpdatesConfiguration.UPDATES_CONFIGURATION_UPDATE_URL_KEY to Uri.parse("https://exp.host/@test/test"),
+        UpdatesConfiguration.UPDATES_CONFIGURATION_CODE_SIGNING_CERTIFICATE to rootCert,
+        UpdatesConfiguration.UPDATES_CONFIGURATION_CODE_SIGNING_METADATA to mapOf(
+          CODE_SIGNING_METADATA_KEY_ID_KEY to "ca-root"
+        ),
+        UpdatesConfiguration.UPDATES_CONFIGURATION_CODE_SIGNING_INCLUDE_MANIFEST_RESPONSE_CERTIFICATE_CHAIN to true,
+      )
+    )
+
+    var errorOccurred: Exception? = null
+    var resultUpdateResponse: UpdateResponse? = null
+
+    FileDownloader(context).parseRemoteUpdateResponse(
+      response, configuration,
+      object : FileDownloader.RemoteUpdateDownloadCallback {
+        override fun onFailure(message: String, e: Exception) {
+          errorOccurred = e
+        }
+
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateResponse = updateResponse
+        }
+      }
+    )
+
+    Assert.assertEquals("Invalid certificate for directive project ID or scope key", errorOccurred!!.message)
+    Assert.assertNull(resultUpdateResponse)
   }
 
   @Test
@@ -498,15 +773,15 @@ class FileDownloaderManifestParsingTest {
     var errorOccurred = false
     var resultUpdateManifest: UpdateManifest? = null
 
-    FileDownloader(context).parseManifestResponse(
+    FileDownloader(context).parseRemoteUpdateResponse(
       response, configuration,
-      object : FileDownloader.ManifestDownloadCallback {
+      object : FileDownloader.RemoteUpdateDownloadCallback {
         override fun onFailure(message: String, e: Exception) {
           errorOccurred = true
         }
 
-        override fun onSuccess(updateManifest: UpdateManifest) {
-          resultUpdateManifest = updateManifest
+        override fun onSuccess(updateResponse: UpdateResponse) {
+          resultUpdateManifest = updateResponse.manifestUpdateResponsePart?.updateManifest
         }
       }
     )