+++ packages/expo-updates/android/src/main/java/expo/modules/updates/loader/FileDownloader.kt
@@ -42,9 +42,9 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
     fun onSuccess(file: File, hash: ByteArray)
   }
 
-  interface ManifestDownloadCallback {
+  interface RemoteUpdateDownloadCallback {
     fun onFailure(message: String, e: Exception)
-    fun onSuccess(updateManifest: UpdateManifest)
+    fun onSuccess(updateResponse: UpdateResponse)
   }
 
   interface AssetDownloadCallback {
@@ -90,13 +90,13 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
     )
   }
 
-  internal fun parseManifestResponse(response: Response, configuration: UpdatesConfiguration, callback: ManifestDownloadCallback) {
+  internal fun parseRemoteUpdateResponse(response: Response, configuration: UpdatesConfiguration, callback: RemoteUpdateDownloadCallback) {
     val contentType = response.header("content-type") ?: ""
     val isMultipart = contentType.startsWith("multipart/", ignoreCase = true)
     if (isMultipart) {
       val boundaryParameter = ParameterParser().parse(contentType, ';')["boundary"]
       if (boundaryParameter == null) {
-        val message = "Missing boundary in multipart manifest content-type"
+        val message = "Missing boundary in multipart remote update content-type"
         logger.error(message, UpdatesErrorCode.UpdateFailedToLoad)
         callback.onFailure(
           message,
@@ -105,18 +105,46 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
         return
       }
 
-      parseMultipartManifestResponse(response, boundaryParameter, configuration, callback)
+      parseMultipartRemoteUpdateResponse(response, boundaryParameter, configuration, callback)
     } else {
       val responseHeaders = response.headers
-      val manifestHeaderData = ManifestHeaderData(
+
+      val responseHeaderData = ResponseHeaderData(
         protocolVersion = responseHeaders["expo-protocol-version"],
-        manifestFilters = responseHeaders["expo-manifest-filters"],
-        serverDefinedHeaders = responseHeaders["expo-server-defined-headers"],
-        manifestSignature = responseHeaders["expo-manifest-signature"],
-        signature = responseHeaders["expo-signature"]
+        manifestFiltersRaw = responseHeaders["expo-manifest-filters"],
+        serverDefinedHeadersRaw = responseHeaders["expo-server-defined-headers"],
+        manifestSignature = responseHeaders["expo-manifest-signature"]
       )
 
-      parseManifest(response.body!!.string(), manifestHeaderData, null, null, configuration, callback)
+      val manifestResponseInfo = ResponsePartInfo(
+        responseHeaderData = responseHeaderData,
+        responsePartHeaderData = ResponsePartHeaderData(
+          signature = responseHeaders["expo-signature"]
+        ),
+        body = response.body!!.string()
+      )
+
+      parseManifest(
+        manifestResponseInfo,
+        null,
+        null,
+        configuration,
+        object : ParseManifestCallback {
+          override fun onFailure(message: String, e: Exception) {
+            callback.onFailure(message, e)
+          }
+
+          override fun onSuccess(manifestUpdateResponsePart: UpdateResponsePart.ManifestUpdateResponsePart) {
+            callback.onSuccess(
+              UpdateResponse(
+                responseHeaderData = responseHeaderData,
+                manifestUpdateResponsePart = manifestUpdateResponsePart,
+                directiveUpdateResponsePart = null
+              )
+            )
+          }
+        }
+      )
     }
   }
 
@@ -135,10 +163,11 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
     return headers.toHeaders()
   }
 
-  private fun parseMultipartManifestResponse(response: Response, boundary: String, configuration: UpdatesConfiguration, callback: ManifestDownloadCallback) {
+  private fun parseMultipartRemoteUpdateResponse(response: Response, boundary: String, configuration: UpdatesConfiguration, callback: RemoteUpdateDownloadCallback) {
     var manifestPartBodyAndHeaders: Pair<String, Headers>? = null
     var extensionsBody: String? = null
     var certificateChainString: String? = null
+    var directivePartBodyAndHeaders: Pair<String, Headers>? = null
 
     val multipartStream = MultipartStream(response.body!!.byteStream(), boundary.toByteArray())
 
@@ -160,13 +189,14 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
               "manifest" -> manifestPartBodyAndHeaders = Pair(output.toString(), headers)
               "extensions" -> extensionsBody = output.toString()
               "certificate_chain" -> certificateChainString = output.toString()
+              "directive" -> directivePartBodyAndHeaders = Pair(output.toString(), headers)
             }
           }
         }
         nextPart = multipartStream.readBoundary()
       }
     } catch (e: Exception) {
-      val message = "Error while reading multipart manifest response"
+      val message = "Error while reading multipart remote update response"
       logger.error(message, UpdatesErrorCode.UpdateFailedToLoad, e)
       callback.onFailure(
         message,
@@ -175,17 +205,10 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
       return
     }
 
-    if (manifestPartBodyAndHeaders == null) {
-      val message = "Multipart manifest response missing manifest part"
-      logger.error(message, UpdatesErrorCode.UpdateFailedToLoad)
-      callback.onFailure(message, IOException("Malformed multipart manifest response"))
-      return
-    }
-
     val extensions = try {
       extensionsBody?.let { JSONObject(it) }
     } catch (e: Exception) {
-      val message = "Failed to parse multipart manifest extensions"
+      val message = "Failed to parse multipart remote update extensions"
       logger.error(message, UpdatesErrorCode.UpdateFailedToLoad)
       callback.onFailure(
         message,
@@ -194,34 +217,196 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
       return
     }
 
+    // in v0 compatibility mode require a manifest
+    if (configuration.enableExpoUpdatesProtocolV0CompatibilityMode && manifestPartBodyAndHeaders == null) {
+      val message = "Multipart response missing manifest part. Manifest is required in version 0 of the expo-updates protocol. This may be due to the update being a rollback or other directive."
+      logger.error(message, UpdatesErrorCode.UpdateFailedToLoad)
+      callback.onFailure(message, IOException(message))
+      return
+    }
+
     val responseHeaders = response.headers
-    val manifestHeaderData = ManifestHeaderData(
+    val responseHeaderData = ResponseHeaderData(
       protocolVersion = responseHeaders["expo-protocol-version"],
-      manifestFilters = responseHeaders["expo-manifest-filters"],
-      serverDefinedHeaders = responseHeaders["expo-server-defined-headers"],
+      manifestFiltersRaw = responseHeaders["expo-manifest-filters"],
+      serverDefinedHeadersRaw = responseHeaders["expo-server-defined-headers"],
       manifestSignature = responseHeaders["expo-manifest-signature"],
-      signature = manifestPartBodyAndHeaders.second["expo-signature"]
     )
 
-    parseManifest(manifestPartBodyAndHeaders.first, manifestHeaderData, extensions, certificateChainString, configuration, callback)
+    val manifestResponseInfo = manifestPartBodyAndHeaders?.let {
+      ResponsePartInfo(
+        responseHeaderData = responseHeaderData,
+        responsePartHeaderData = ResponsePartHeaderData(
+          signature = manifestPartBodyAndHeaders.second["expo-signature"]
+        ),
+        body = manifestPartBodyAndHeaders.first
+      )
+    }
+
+    // in v0 compatibility mode ignore directives
+    val directiveResponseInfo = if (configuration.enableExpoUpdatesProtocolV0CompatibilityMode) {
+      null
+    } else {
+      directivePartBodyAndHeaders?.let {
+        ResponsePartInfo(
+          responseHeaderData = responseHeaderData,
+          responsePartHeaderData = ResponsePartHeaderData(
+            signature = directivePartBodyAndHeaders.second["expo-signature"]
+          ),
+          body = directivePartBodyAndHeaders.first
+        )
+      }
+    }
+
+    var parseManifestResponse: UpdateResponsePart.ManifestUpdateResponsePart? = null
+    var parseDirectiveResponse: UpdateResponsePart.DirectiveUpdateResponsePart? = null
+    var didError = false
+
+    // need to parse the directive and manifest in parallel, to do so use this common callback.
+    // would be a great place to have better coroutine stuff
+    val maybeFinish = {
+      if (!didError) {
+        val isManifestDone = manifestResponseInfo == null || parseManifestResponse != null
+        val isDirectiveDone = directiveResponseInfo == null || parseDirectiveResponse != null
+
+        if (isManifestDone && isDirectiveDone) {
+          callback.onSuccess(
+            UpdateResponse(
+              responseHeaderData = responseHeaderData,
+              manifestUpdateResponsePart = parseManifestResponse,
+              directiveUpdateResponsePart = parseDirectiveResponse
+            )
+          )
+        }
+      }
+    }
+
+    if (directiveResponseInfo != null) {
+      parseDirective(
+        directiveResponseInfo,
+        certificateChainString,
+        configuration,
+        object : ParseDirectiveCallback {
+          override fun onFailure(message: String, e: Exception) {
+            if (!didError) {
+              didError = true
+              callback.onFailure(message, e)
+            }
+          }
+
+          override fun onSuccess(directiveUpdateResponsePart: UpdateResponsePart.DirectiveUpdateResponsePart) {
+            parseDirectiveResponse = directiveUpdateResponsePart
+            maybeFinish()
+          }
+        }
+      )
+    }
+
+    if (manifestResponseInfo != null) {
+      parseManifest(
+        manifestResponseInfo,
+        extensions,
+        certificateChainString,
+        configuration,
+        object : ParseManifestCallback {
+          override fun onFailure(message: String, e: Exception) {
+            if (!didError) {
+              didError = true
+              callback.onFailure(message, e)
+            }
+          }
+          override fun onSuccess(manifestUpdateResponsePart: UpdateResponsePart.ManifestUpdateResponsePart) {
+            parseManifestResponse = manifestUpdateResponsePart
+            maybeFinish()
+          }
+        }
+      )
+    }
+
+    // if both parts are empty, we still want to finish
+    if (manifestResponseInfo == null && directiveResponseInfo == null) {
+      maybeFinish()
+    }
+  }
+
+  interface ParseDirectiveCallback {
+    fun onFailure(message: String, e: Exception)
+    fun onSuccess(directiveUpdateResponsePart: UpdateResponsePart.DirectiveUpdateResponsePart)
+  }
+
+  private fun parseDirective(
+    directiveResponsePartInfo: ResponsePartInfo,
+    certificateChainFromManifestResponse: String?,
+    configuration: UpdatesConfiguration,
+    callback: ParseDirectiveCallback
+  ) {
+    try {
+      val body = directiveResponsePartInfo.body
+
+      // check code signing if code signing is configured
+      // 1. verify the code signing signature (throw if invalid)
+      // 2. then, if the code signing certificate is only valid for a particular project, verify that the directive
+      //    has the correct info for code signing. If the code signing certificate doesn't specify a particular
+      //    project, it is assumed to be valid for all projects
+      // 3. consider the directive verified if both of these pass
+      try {
+        configuration.codeSigningConfiguration?.let { codeSigningConfiguration ->
+          val signatureValidationResult = codeSigningConfiguration.validateSignature(
+            directiveResponsePartInfo.responsePartHeaderData.signature,
+            body.toByteArray(),
+            certificateChainFromManifestResponse,
+          )
+          if (signatureValidationResult.validationResult == ValidationResult.INVALID) {
+            throw IOException("Directive download was successful, but signature was incorrect")
+          }
+
+          if (signatureValidationResult.validationResult != ValidationResult.SKIPPED) {
+            val directiveForProjectInformation = UpdateDirective.fromJSONString(body)
+            signatureValidationResult.expoProjectInformation?.let { expoProjectInformation ->
+              if (expoProjectInformation.projectId != directiveForProjectInformation.signingInfo?.easProjectId ||
+                expoProjectInformation.scopeKey != directiveForProjectInformation.signingInfo.scopeKey
+              ) {
+                throw CertificateException("Invalid certificate for directive project ID or scope key")
+              }
+            }
+          }
+        }
+      } catch (e: Exception) {
+        callback.onFailure(e.message!!, e)
+        return
+      }
+
+      callback.onSuccess(UpdateResponsePart.DirectiveUpdateResponsePart(UpdateDirective.fromJSONString(body)))
+    } catch (e: Exception) {
+      val message = "Failed to parse directive data: ${e.localizedMessage}"
+      logger.error(message, UpdatesErrorCode.UpdateFailedToLoad, e)
+      callback.onFailure(
+        message,
+        e
+      )
+    }
+  }
+
+  interface ParseManifestCallback {
+    fun onFailure(message: String, e: Exception)
+    fun onSuccess(manifestUpdateResponsePart: UpdateResponsePart.ManifestUpdateResponsePart)
   }
 
   private fun parseManifest(
-    manifestBody: String,
-    manifestHeaderData: ManifestHeaderData,
+    manifestResponseInfo: ResponsePartInfo,
     extensions: JSONObject?,
     certificateChainFromManifestResponse: String?,
     configuration: UpdatesConfiguration,
-    callback: ManifestDownloadCallback
+    callback: ParseManifestCallback
   ) {
     try {
-      val updateResponseJson = extractUpdateResponseJson(manifestBody, configuration)
+      val updateResponseJson = extractUpdateResponseJson(manifestResponseInfo.body, configuration)
       val isSignatureInBody =
         updateResponseJson.has("manifestString") && updateResponseJson.has("signature")
       val signature = if (isSignatureInBody) {
         updateResponseJson.getNullable("signature")
       } else {
-        manifestHeaderData.manifestSignature
+        manifestResponseInfo.responseHeaderData.manifestSignature
       }
 
       /**
@@ -235,7 +420,7 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
       val manifestString = if (isSignatureInBody) {
         updateResponseJson.getString("manifestString")
       } else {
-        manifestBody
+        manifestResponseInfo.body
       }
       val preManifest = JSONObject(manifestString)
 
@@ -255,7 +440,18 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
             override fun onCompleted(isValid: Boolean) {
               if (isValid) {
                 try {
-                  checkCodeSigningAndCreateManifest(manifestBody, preManifest, manifestHeaderData, extensions, certificateChainFromManifestResponse, true, configuration, logger, callback)
+                  checkCodeSigningAndCreateManifest(
+                    bodyString = manifestResponseInfo.body,
+                    preManifest = preManifest,
+                    responseHeaderData = manifestResponseInfo.responseHeaderData,
+                    responsePartHeaderData = manifestResponseInfo.responsePartHeaderData,
+                    extensions = extensions,
+                    certificateChainFromManifestResponse = certificateChainFromManifestResponse,
+                    isVerified = true,
+                    configuration = configuration,
+                    logger = logger,
+                    callback = callback
+                  )
                 } catch (e: Exception) {
                   callback.onFailure("Failed to parse manifest data", e)
                 }
@@ -271,7 +467,18 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
           }
         )
       } else {
-        checkCodeSigningAndCreateManifest(manifestBody, preManifest, manifestHeaderData, extensions, certificateChainFromManifestResponse, false, configuration, logger, callback)
+        checkCodeSigningAndCreateManifest(
+          bodyString = manifestResponseInfo.body,
+          preManifest = preManifest,
+          responseHeaderData = manifestResponseInfo.responseHeaderData,
+          responsePartHeaderData = manifestResponseInfo.responsePartHeaderData,
+          extensions = extensions,
+          certificateChainFromManifestResponse = certificateChainFromManifestResponse,
+          isVerified = false,
+          configuration = configuration,
+          logger = logger,
+          callback = callback
+        )
       }
     } catch (e: Exception) {
       val message = "Failed to parse manifest data: ${e.localizedMessage}"
@@ -283,18 +490,18 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
     }
   }
 
-  fun downloadManifest(
+  fun downloadRemoteUpdate(
     configuration: UpdatesConfiguration,
     extraHeaders: JSONObject?,
     context: Context,
-    callback: ManifestDownloadCallback
+    callback: RemoteUpdateDownloadCallback
   ) {
     try {
       downloadData(
-        createRequestForManifest(configuration, extraHeaders, context),
+        createRequestForRemoteUpdate(configuration, extraHeaders, context),
         object : Callback {
           override fun onFailure(call: Call, e: IOException) {
-            val message = "Failed to download manifest from URL: ${configuration.updateUrl}: ${e.localizedMessage}"
+            val message = "Failed to download remote update from URL: ${configuration.updateUrl}: ${e.localizedMessage}"
             logger.error(message, UpdatesErrorCode.UpdateFailedToLoad, e)
             callback.onFailure(
               message,
@@ -305,23 +512,21 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
           @Throws(IOException::class)
           override fun onResponse(call: Call, response: Response) {
             if (!response.isSuccessful) {
-              val message = "Failed to download manifest from URL: ${configuration.updateUrl}"
+              val message = "Failed to download remote update from URL: ${configuration.updateUrl}"
               logger.error(message, UpdatesErrorCode.UpdateFailedToLoad)
               callback.onFailure(
                 message,
-                Exception(
-                  response.body!!.string()
-                )
+                Exception(response.body!!.string())
               )
               return
             }
 
-            parseManifestResponse(response, configuration, callback)
+            parseRemoteUpdateResponse(response, configuration, callback)
           }
         }
       )
     } catch (e: Exception) {
-      val message = "Failed to download manifest from URL: ${configuration.updateUrl}: ${e.localizedMessage}"
+      val message = "Failed to download remote update from URL: ${configuration.updateUrl}: ${e.localizedMessage}"
       logger.error(message, UpdatesErrorCode.UpdateFailedToLoad, e)
       callback.onFailure(
         message,
@@ -405,13 +610,14 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
     private fun checkCodeSigningAndCreateManifest(
       bodyString: String,
       preManifest: JSONObject,
-      manifestHeaderData: ManifestHeaderData,
+      responseHeaderData: ResponseHeaderData,
+      responsePartHeaderData: ResponsePartHeaderData,
       extensions: JSONObject?,
       certificateChainFromManifestResponse: String?,
       isVerified: Boolean,
       configuration: UpdatesConfiguration,
       logger: UpdatesLogger,
-      callback: ManifestDownloadCallback
+      callback: ParseManifestCallback
     ) {
       if (configuration.expectsSignedManifest) {
         preManifest.put("isVerified", isVerified)
@@ -426,7 +632,7 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
       try {
         configuration.codeSigningConfiguration?.let { codeSigningConfiguration ->
           val signatureValidationResult = codeSigningConfiguration.validateSignature(
-            manifestHeaderData.signature,
+            responsePartHeaderData.signature,
             bodyString.toByteArray(),
             certificateChainFromManifestResponse,
           )
@@ -437,7 +643,7 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
           if (signatureValidationResult.validationResult != ValidationResult.SKIPPED) {
             val manifestForProjectInformation = ManifestFactory.getManifest(
               preManifest,
-              manifestHeaderData,
+              responseHeaderData,
               extensions,
               configuration
             ).manifest
@@ -459,13 +665,13 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
         return
       }
 
-      val updateManifest = ManifestFactory.getManifest(preManifest, manifestHeaderData, extensions, configuration)
-      if (!SelectionPolicies.matchesFilters(updateManifest.updateEntity!!, updateManifest.manifestFilters)) {
+      val updateManifest = ManifestFactory.getManifest(preManifest, responseHeaderData, extensions, configuration)
+      if (!SelectionPolicies.matchesFilters(updateManifest.updateEntity!!, responseHeaderData.manifestFilters)) {
         val message =
           "Downloaded manifest is invalid; provides filters that do not match its content"
         callback.onFailure(message, Exception(message))
       } else {
-        callback.onSuccess(updateManifest)
+        callback.onSuccess(UpdateResponsePart.ManifestUpdateResponsePart(updateManifest))
       }
     }
 
@@ -522,6 +728,7 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
         .url(assetEntity.url!!.toString())
         .addHeadersFromJSONObject(assetEntity.extraRequestHeaders)
         .header("Expo-Platform", "android")
+        .header("Expo-Protocol-Version", "1")
         .header("Expo-API-Version", "1")
         .header("Expo-Updates-Environment", "BARE")
         .header("EAS-Client-ID", EASClientID(context).uuid.toString())
@@ -533,7 +740,7 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
         .build()
     }
 
-    internal fun createRequestForManifest(
+    internal fun createRequestForRemoteUpdate(
       configuration: UpdatesConfiguration,
       extraHeaders: JSONObject?,
       context: Context
@@ -543,6 +750,7 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
         .addHeadersFromJSONObject(extraHeaders)
         .header("Accept", "multipart/mixed,application/expo+json,application/json")
         .header("Expo-Platform", "android")
+        .header("Expo-Protocol-Version", "1")
         .header("Expo-API-Version", "1")
         .header("Expo-Updates-Environment", "BARE")
         .header("Expo-JSON-Error", "true")