+++ packages/expo-updates/src/Updates.types.ts
@@ -35,6 +35,13 @@ export type Manifest = ClassicManifest | typeof Constants.manifest2;
 // modern manifest type is intentionally not exported, since the plan is to call it just "Manifest"
 // in the future
 
+type UpdateCheckResultRollBackToEmbedded = {
+  /**
+   * Signifies that a roll back update is available.
+   */
+  isRollBackToEmbedded: true;
+};
+
 /**
  * The successful result of checking for a new update.
  */
@@ -66,7 +73,10 @@ type UpdateCheckResultFailure = {
 /**
  * The result of checking for a new update.
  */
-export type UpdateCheckResult = UpdateCheckResultSuccess | UpdateCheckResultFailure;
+export type UpdateCheckResult =
+  | UpdateCheckResultRollBackToEmbedded
+  | UpdateCheckResultSuccess
+  | UpdateCheckResultFailure;
 
 /**
  * The successful result of fetching a new update.
@@ -98,9 +108,22 @@ export type UpdateFetchResultFailure = {
 };
 
 /**
+ * The rollback to embedded result of fetching a new update.
+ */
+type UpdateFetchResultRollbackToEmbedded = {
+  /**
+   * Signifies that the update was a roll back to the embedded update.
+   */
+  isRollBackToEmbedded: true;
+};
+
+/**
  * The result of fetching a new update.
  */
-export type UpdateFetchResult = UpdateFetchResultSuccess | UpdateFetchResultFailure;
+export type UpdateFetchResult =
+  | UpdateFetchResultSuccess
+  | UpdateFetchResultFailure
+  | UpdateFetchResultRollbackToEmbedded;
 
 /**
  * An object that is passed into each event listener when an auto-update check occurs.